pyproject.toml
[project]
name = "Yantra"
version = "1.0.0"
description = "Common config for mlops"
dependencies = [
    "PyYAML==6.0.2",
    "dvc-s3 == 3.0.1",
    "mlflow >= 2.10.0",
    "prefect >= 2.14.0",
    "evidently >= 0.4.0",
    "boto3 >= 1.34.0"
]


[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]